# docs/swagger.yaml
definitions:
  global.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      traceId:
        type: string
    type: object
  validate.Login:
    properties:
      captchaId:
        type: string
      code:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captchaId
    - code
    - password
    - username
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: 25076778@qq.com
    name: API Support
    url: http://www.swagger.io/support
  description: Gin-base API 文档
  termsOfService: http://swagger.io/terms/
  title: Gin-base Swagger API
  version: "2.0"
paths:
...
  /api/v1/captcha:
    get:
      consumes:
      - application/json
      description: 获取验证码
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 获取验证码
      tags:
      - 登录相关
    post:
      description: 校验验证码
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 校验验证码
      tags:
      - 登录相关
  /api/v1/dict:
    get:
      description: 字典列表
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 列表
      tags:
      - 字典管理
    post:
      description: 字典创建
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 创建
      tags:
      - 字典管理
  /api/v1/dict/{id}:
    delete:
      description: 字典删除
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 删除
      tags:
      - 字典管理
    get:
      description: 字典详情
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 详情
      tags:
      - 字典管理
    put:
      description: 字典更新
      parameters:
      - description: 认证Token
        in: header
        name: token
        required: true
        type: string
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 更新
      tags:
      - 字典管理
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/validate.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回" Example({"code":0,"msg":"Success","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "400":
          description: 参数错误" Example({"code":400,"msg":"参数错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
        "500":
          description: 系统错误" Example({"code":500,"msg":"系统错误","data":[]})
          schema:
            $ref: '#/definitions/global.Response'
      summary: 登录
      tags:
      - 登录相关
...
swagger: "2.0"